+++ examples/basic-i18n/src/routes.ts
@@ -1,0 +1,25 @@
+import { IRouterConfig } from 'ice';
+import BasicLayout from '@/layouts/BasicLayout';
+import Home from '@/pages/Home';
+import About from '@/pages/About';
+
+const routerConfig: IRouterConfig[] = [
+  {
+    path: '/',
+    component: BasicLayout,
+    children: [
+      {
+        path: '/about',
+        exact: true,
+        component: About
+      },
+      {
+        path: '/',
+        exact: true,
+        component: Home,
+      },
+    ]
+  },
+];
+
+export default routerConfig;