+++ examples/basic-i18n/src/typing.d.ts
@@ -1,0 +1,6 @@
+declare module '*.module.less' {
+  const classes: { [key: string]: string };
+  export default classes;
+}
+
+declare module '*.svg'