+++ packages/plugin-i18n/src/constants.ts
@@ -1,0 +1,1 @@
+export const LOCALE_COOKIE_KEY = 'ice_locale';