+++ packages/vite-plugin-index-html/tests/utils.ts
@@ -1,18 +1,16 @@
-import { getTheFirstEntry, isSingleEntry } from '../src/utils';
+import { isSingleEntry } from '../src/utils';
 
 describe('vite-plugin-index-html-utils', () => {
   test('entry', () => {
     let entry: string | Record<string, string> = './src/main.tsx';
 
     expect(isSingleEntry(entry)).toBeTruthy();
-    expect(getTheFirstEntry(entry)).toBe('./src/main.tsx');
 
     entry = {
       index: './src/main.tsx'
     }
 
     expect(isSingleEntry(entry)).toBeTruthy();
-    expect(getTheFirstEntry(entry)).toBe('./src/main.tsx');
 
     entry = {
       index: './src/main.tsx',