+++ packages/vite-service/src/start.ts
@@ -50,7 +50,7 @@ export async function viteStart(context: Context): Promise<StartResult> {
 function generateDevServerUrl(devConfig) {
   const { server: { port, https, host } } = devConfig;
   const protocol = https ? 'https' : 'http';
-  return `${protocol}://${resolveHostname(host)}:${port}`;
+  return `${protocol}://${resolveHostname(host)}:${port}/`;
 }
 
 // Reference: https://github.com/vitejs/vite/blob/7e3e84e1b733f4cb0cba3bd69f28a5671b52261c/packages/vite/src/node/utils.ts#L580