+++ packages/build-app-helpers/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## 2.5.0
 
+- [chore] revert `@babel/plugin-transform-runtime` helpers value
 - [chore] bump version of `es-module-lexer`
 - [feat] support scan source code of import statement
+
+
+## 2.4.3
+
+- [chore] not inline @babel/runtime/helpers code
 
 ## 2.4.2