+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.0
+
+- [feat] support `--force` to empty cache folder
+
+## 2.0.5
+
+- [fix] `process.env.__IS_SERVER__` default value is `false`
+
 ## 2.0.4
 
 - [fix] `minify: true` is invalid