+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## 3.0.7
+## 3.1.0
 
-- [chore] move runtime's dependencies to `@ice/runtime`
+- [feat] support SSR in mode Vite
 
 ## 3.0.6